cleaning in embObjLib: separate files for complex classes, using the eth:: namespace
