+++ src/libraries/icubmod/embObjAnalog/embObjAnalogSensor.h
@@ -20,16 +20,11 @@
 
 #include <iCub/FactoryInterface.h>
 #include <iCub/LoggerInterfaces.h>
-#include <hostTransceiver.hpp>
-#include <ethResource.h>
-#include <ethManager.h>
-
-
-#include "FeatureInterface.h"  
-
-#include "EoAnalogSensors.h"
 
 #include "IethResource.h"
+#include <ethManager.h>
+#include <abstractEthResource.h>
+
 
 #include <yarp/os/LogStream.h>
 
@@ -37,7 +32,6 @@
 namespace yarp{
     namespace dev{
         class embObjAnalogSensor;
-        class TheEthManager;
     }
 }
 
@@ -78,7 +72,7 @@ typedef int AnalogDataFormat;
  */
 class yarp::dev::embObjAnalogSensor:    public yarp::dev::IAnalogSensor,
                                         public yarp::dev::DeviceDriver,
-                                        public IethResource
+                                        public eth::IethResource
 {
 
 public:
@@ -114,7 +108,7 @@ public:
 
     // IethResource interface
     virtual bool initialised();
-    virtual iethresType_t type();
+    virtual eth::iethresType_t type();
     virtual bool update(eOprotID32_t id32, double timestamp, void* rxdata);
 
 private:
@@ -122,8 +116,8 @@ private:
     char boardIPstring[20];
 
     //! eth messaging stuff
-    TheEthManager       *ethManager;
-    AbstractEthResource *res;
+    eth::TheEthManager *ethManager;
+    eth::AbstractEthResource *res;
 
     bool opened;
     bool verbosewhenok;
