+++ src/libraries/icubmod/embObjLib/ethManager.h
@@ -1,14 +1,10 @@
-
 // -*- mode:C++; tab-width:4; c-basic-offset:4; indent-tabs-mode:nil -*-
 
-/**
- * @ingroup icub_hardware_modules
- * \defgroup TheEthManager TheEthManager
- *
-*/
-/* Copyright (C) 2012  iCub Facility, Istituto Italiano di Tecnologia
- * Author: Alberto Cardellino
- * email: alberto.cardellino@iit.it
+/*
+ * Copyright (C) 2017 iCub Facility - Istituto Italiano di Tecnologia
+ * Author:  Alberto Cardellino, Marco Accame
+ * email:   alberto.cardellino@iit.it, marco.accame@iit.it
+ * website: www.robotcub.org
  * Permission is granted to copy, distribute, and/or modify this program
  * under the terms of the GNU General Public License, version 2 or any
  * later version published by the Free Software Foundation.
@@ -20,15 +16,12 @@
  * WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
  * Public License for more details
- */
+*/
 
-//
-// $Id: TheEthManager.h,v 1.5 2008/06/25 22:33:53 nat Exp $
-//
-//
+// - include guard ----------------------------------------------------------------------------------------------------
 
-#ifndef __ethManager__
-#define __ethManager__
+#ifndef _THEETHMANAGER_H_
+#define _THEETHMANAGER_H_
 
 // Standard includes
 #include <iostream>
@@ -36,7 +29,7 @@
 #include <list>
 #include <string>
 #include <stdio.h>
-#include <map>
+//#include <map>
 
 
 // ACE includes
@@ -61,7 +54,7 @@
 #include "hostTransceiver.hpp"
 #include "FeatureInterface.h"
 #include "IethResource.h"
-#include "ethResource.h"
+#include "abstractEthResource.h"
 
 // embobj includes
 #include "EoProtocol.h"
@@ -76,88 +69,11 @@ using namespace yarp::dev;
 using namespace std;
 
 
-// -- class EthBoards
-// -- it collects all the ETH boards managed by ethManager.
-// -- each board surely has an EthResource object associated to it. and it may have one or more interfaces which use the
-// -- services of EthResource to transmit or receive.
-// -- it is responsibility of the object which owns EthBoards (it is ethManager) to protect the class EthBoards vs concurrent use.
-// -- examples of concurrent use are: transmit or receive using an ethresource and ... attempting to create or destroy a resource.
-
-typedef struct
-{
-    IethResource* interface;
-    iethresType_t type;
-} interfaceInfo_t;
-
-class EthBoards
-{
-
-public:
-
-    enum { maxEthBoards = 32 };
-
-public:
-
-    EthBoards();
-    ~EthBoards();
-
-    size_t number_of_resources(void);
-    bool add(AbstractEthResource* res);
-    AbstractEthResource* get_resource(eOipv4addr_t ipv4);
-    bool rem(AbstractEthResource* res);
-
-    size_t number_of_interfaces(AbstractEthResource* res);
-    bool add(AbstractEthResource* res, IethResource* interface);
-    IethResource* get_interface(eOipv4addr_t ipv4, eOprotID32_t id32);
-    IethResource* get_interface(eOipv4addr_t ipv4, iethresType_t type);
-    bool rem(AbstractEthResource* res, iethresType_t type);
-    
-
-    // the name of the board
-    const char * name(eOipv4addr_t ipv4);
-
-    // executes an action on all EthResource which have been added in the class.
-    bool execute(void (*action)(AbstractEthResource* res, void* p), void* par);
-
-    // executes an action on the ethResource having a specific ipv4.
-    bool execute(eOipv4addr_t ipv4, void (*action)(AbstractEthResource* res, void* p), void* par);
 
 
-private:
-
-    // private types
-
-    enum { ethboardNameMaxSize = EthResource::boardNameSize };
-
-    typedef struct
-    {
-        eOipv4addr_t        ipv4;
-        eObrd_ethtype_t     type;
-        const char *        nameoftype;
-        char                name[EthBoards::ethboardNameMaxSize];
-        uint8_t             numberofinterfaces;
-        uint8_t             boardnumber;
-        AbstractEthResource*  resource;
-        IethResource*       interfaces[iethresType_numberof];
-    } ethboardProperties_t;
-
-
-private:
-
-    // private variables
-
-    static const char * defaultnames[EthBoards::maxEthBoards];
-    static const char * errorname[1];
-
-    int sizeofLUT;
-    ethboardProperties_t LUT[EthBoards::maxEthBoards];
-
-private:
-
-    // private functions
-    bool get_LUTindex(eOipv4addr_t ipv4, uint8_t &index);
-};
-
+#include <ethBoards.h>
+#include <ethSender.h>
+#include <ethReceiver.h>
 
 
 // -- class TheEthManager
@@ -166,171 +82,115 @@ private:
 // -- it holds class EthBoards which stores references to EthResource (what is used to form / parse UDP packets for the eth board)
 // -- and to the interfaces which use the EthResource of a given board.
 
-// forward declaration because they are used inside TheEthManager
-class EthSender;
-class EthReceiver;
+namespace eth {
 
-class yarp::dev::TheEthManager: public DeviceDriver
-{
+    class TheEthManager
+    {
 
-public:
-    // this is the maximum number of boards that the singleton can manage.
-    enum { maxBoards = EthBoards::maxEthBoards };
+    public:
+        // this is the maximum number of boards that the singleton can manage.
+        enum { maxBoards = eth::EthBoards::maxEthBoards };
 
-    // these are the boards, their use is protected by txSem or rxSem or both of them.
-    EthBoards* ethBoards;
+        // these are the boards, their use is protected by txSem or rxSem or both of them.
+        eth::EthBoards* ethBoards;
 
-private:
+    private:
 
-    // singletons have private constructor / destructor
-    TheEthManager();
-    ~TheEthManager();
+        // singletons have private constructor / destructor
+        TheEthManager();
+        ~TheEthManager();
 
-public:
+    public:
 
-    // singletons use this method to retrieve an handle to the object
-    static TheEthManager* instance();
+        // singletons use this method to retrieve an handle to the object
+        static TheEthManager* instance();
 
-    // we use this method to for a destruction of the object.
-    static bool killYourself();
-
-    bool open(void);
-    bool close(void);
-
-    // useful for printing times relative to start of the object
-    double getLifeTime(void);
-
-    bool verifyEthBoardInfo(yarp::os::Searchable &cfgtotal, eOipv4addr_t* boardipv4, char *boardipv4string, int stringsize, char *boardNameStr=NULL, int sizeofBoardNameStr=0);
+        // we use this method to for a destruction of the object.
+        static bool killYourself();
 
 
-    AbstractEthResource* requestResource2(IethResource *interface, yarp::os::Searchable &cfgtotal);
+        // useful for printing times relative to start of the object
+        double getLifeTime(void);
 
-    int releaseResource2(AbstractEthResource* ethresource, IethResource* interface);
-
-    const ACE_INET_Addr& getLocalIPaddress(void);
-
-    const eOipv4addressing_t & getLocalIPV4addressing(void);
-
-    bool Transmission(void);
-
-    bool CheckPresence(void);
-
-    bool Reception(ACE_INET_Addr adr, uint64_t* data, ssize_t size, bool collectStatistics);
-
-    AbstractEthResource* getEthResource(eOipv4addr_t ipv4);
-
-    IethResource* getInterface(eOipv4addr_t ipv4, eOprotID32_t id32);
-
-    int getNumberOfResources(void);
-
-    const char * getName(eOipv4addr_t ipv4);
-
-    int sendPacket(void *udpframe, size_t len, ACE_INET_Addr toaddress);
-
-private:
-
-    bool isCommunicationInitted(void);
-
-    bool createCommunicationObjects(ACE_INET_Addr local_addr, int txrate, int rxrate);
-
-    bool initCommunication(yarp::os::Searchable &cfgtotal);
-
-    bool stopCommunicationThreads(void);
-
-    bool lock(bool on);
-
-    bool lockTX(bool on);
-    bool lockRX(bool on);
-    bool lockTXRX(bool on);
+        bool verifyEthBoardInfo(yarp::os::Searchable &cfgtotal, eOipv4addr_t* boardipv4, char *boardipv4string, int stringsize, char *boardNameStr=NULL, int sizeofBoardNameStr=0);
 
 
-private:
+        eth::AbstractEthResource* requestResource2(IethResource *interface, yarp::os::Searchable &cfgtotal);
 
-    void initEOYsystem(void);
+        int releaseResource2(eth::AbstractEthResource* ethresource, IethResource* interface);
 
-    // this semaphore is used to ....
-    static yarp::os::Semaphore managerSem;
-    // the following two semaphore are used separately or together to stop tx and rx if a change is done on ethboards (in startup and shutdown phases)
-    static yarp::os::Semaphore txSem;
-    static yarp::os::Semaphore rxSem;
+        const ACE_INET_Addr& getLocalIPaddress(void);
 
-    static TheEthManager* handle;
+        const eOipv4addressing_t & getLocalIPV4addressing(void);
 
-    // contains the start-up time of the system so that time measures / prints can be done relative.
-    double startUpTime;
+        bool Transmission(void);
 
-    ACE_INET_Addr localIPaddress;
-    eOipv4addressing_t ipv4local;
+        bool CheckPresence(void);
 
-    bool communicationIsInitted;
+        bool Reception(ACE_INET_Addr adr, uint64_t* data, ssize_t size, bool collectStatistics);
 
-    // periodic threads which use methods of class TheEthManager to transmit / receive + the udp socket
-    EthSender* sender;
-    EthReceiver* receiver;
-    ACE_SOCK_Dgram* UDP_socket;
-    bool embBoardsConnected;
+        eth::AbstractEthResource* getEthResource(eOipv4addr_t ipv4);
 
-};
+        IethResource* getInterface(eOipv4addr_t ipv4, eOprotID32_t id32);
 
+        int getNumberOfResources(void);
 
-// -- class EthSender
-// -- it is a rate thread created by singleton TheEthManager. it regularly transmits packets (if any available) to the eth boards.
-// -- it uses methods made available by TheEthManager.
+        const char * getName(eOipv4addr_t ipv4);
 
-class EthSender : public yarp::os::RateThread
-{
-private:
-    int rateofthread;
+        int sendPacket(void *udpframe, size_t len, ACE_INET_Addr toaddress);
 
-    uint8_t                       *p_sendData;
-    TheEthManager                 *ethManager;
-    ACE_SOCK_Dgram                *send_socket;
-    void run();
+    private:
+
+        bool isCommunicationInitted(void);
+
+        bool createCommunicationObjects(ACE_INET_Addr local_addr, int txrate, int rxrate);
+
+        bool initCommunication(yarp::os::Searchable &cfgtotal);
+
+        bool stopCommunicationThreads(void);
+
+        bool lock(bool on);
+
+        bool lockTX(bool on);
+        bool lockRX(bool on);
+        bool lockTXRX(bool on);
 
 
-public:
+    private:
 
-    enum { EthSenderDefaultRate = 1, EthSenderMaxRate = 20 };
+        void initEOYsystem(void);
 
-    EthSender(int txrate);
-    ~EthSender();
-    bool config(ACE_SOCK_Dgram *pSocket, TheEthManager* _ethManager);
-    bool threadInit();
+        // this semaphore is used to ....
+        static yarp::os::Semaphore managerSem;
+        // the following two semaphore are used separately or together to stop tx and rx if a change is done on ethboards (in startup and shutdown phases)
+        static yarp::os::Semaphore txSem;
+        static yarp::os::Semaphore rxSem;
 
-};
+        static eth::TheEthManager* handle;
 
+        // contains the start-up time of the system so that time measures / prints can be done relative.
+        double startUpTime;
 
-// -- class EthReceiver
-// -- it is a rate thread created by singleton TheEthManager.
-// -- it regularly wakes up to see if a packet is in its listening socket and it parses that with methods made available by TheEthManager.
+        ACE_INET_Addr localIPaddress;
+        eOipv4addressing_t ipv4local;
 
-class EthReceiver : public yarp::os::RateThread
-{
-private:
-    int rateofthread;
+        bool communicationIsInitted;
 
-    ACE_SOCK_Dgram                  *recv_socket;
-    TheEthManager                   *ethManager;
-    double                          statPrintInterval;
+        // periodic threads which use methods of class TheEthManager to transmit / receive + the udp socket
+        eth::EthSender* sender;
+        eth::EthReceiver* receiver;
+        ACE_SOCK_Dgram* UDP_socket;
+        bool embBoardsConnected;
 
+    };
 
-public:
-
-    enum { EthReceiverDefaultRate = 5, EthReceiverMaxRate = 20 };
-
-    EthReceiver(int rxrate);
-    ~EthReceiver();
-    bool config(ACE_SOCK_Dgram *pSocket, TheEthManager* _ethManager);
-    bool threadInit();
-    void run();
-    void onStop();
-};
+} // namespace eth
 
 
+#endif  // include-guard
 
-#endif
 
-// eof
+// - end-of-file (leave a blank line after)----------------------------------------------------------------------------
