+++ src/libraries/icubmod/embObjMais/embObjMais.h
@@ -20,15 +20,12 @@
 
 #include <iCub/FactoryInterface.h>
 #include <iCub/LoggerInterfaces.h>
-#include <hostTransceiver.hpp>
-#include <ethResource.h>
-#include <ethManager.h>
 
-
-#include "FeatureInterface.h"  
-#include "EoAnalogSensors.h"
 
 #include "IethResource.h"
+#include <ethManager.h>
+#include <abstractEthResource.h>
+
 
 #include <yarp/os/LogStream.h>
 
@@ -37,7 +34,6 @@
 namespace yarp {
     namespace dev {
         class embObjMais;
-        class TheEthManager;
     }
 }
 
@@ -49,7 +45,7 @@ namespace yarp {
 
 class yarp::dev::embObjMais:            public yarp::dev::IAnalogSensor,
                                         public yarp::dev::DeviceDriver,
-                                        public IethResource
+                                        public eth::IethResource
 {
 
 public:
@@ -76,15 +72,15 @@ public:
 
     // IethResource interface
     virtual bool initialised();
-    virtual iethresType_t type();
+    virtual eth::iethresType_t type();
     virtual bool update(eOprotID32_t id32, double timestamp, void* rxdata);
 
 private:
 
     char boardIPstring[20];
 
-    TheEthManager* ethManager;
-    AbstractEthResource* res;
+    eth::TheEthManager* ethManager;
+    eth::AbstractEthResource* res;
     ServiceParser* parser;
 
     bool opened;
